double width of sUnassociatedStationStats.time_stamp to delay overflow
EasyMesh Unassociated STA Link Metrics Response TLV contains
a 4byte / 32bit field that should hold the time delta
between the 'time point when the measurement was made' and
the 'time point when the Report was sent'

PWHM exposes time as a RFC3339 string
using lib amxc to parse this string into a struct containing
a signed 64bit 'seconds' field.

+++ common/beerocks/bwl/include/bwl/base_wlan_hal_types.h
@@ -224,7 +224,7 @@ typedef struct {
     uint8_t signal_strength;
     uint8_t channel;
     uint8_t operating_class;
-    uint32_t time_stamp;
+    uint64_t time_stamp;
 } sUnassociatedStationStats;
 
 typedef struct {
